--- src/main/java/com/android/volley/toolbox/ThrowingCache.java
@@ -1,54 +1,0 @@
-/*
- * Copyright (C) 2020 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.android.volley.toolbox;
-
-import androidx.annotation.RestrictTo;
-import com.android.volley.Cache;
-
-/** A cache that throws an error if a method is called. */
-@RestrictTo(RestrictTo.Scope.LIBRARY)
-public class ThrowingCache implements Cache {
-    @Override
-    public Entry get(String key) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void put(String key, Entry entry) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void initialize() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void invalidate(String key, boolean fullExpire) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void remove(String key) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void clear() {
-        throw new UnsupportedOperationException();
-    }
-}
